add wanaazegaj ekid and every navgation

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -17,7 +17,7 @@ class CreateUsersTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('username')->unique();
-            $table->integer('salary')->nullable();
+            $table->decimal('salary')->nullable();
             $table->string('upload_image')->nullable();
             $table->integer('role_id');
             $table->integer('status_id')->nullable(); // may be unneccesary
